spatial-cell-interactions/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spatial-cell-interactions"
version = "0.1.0"
description = "Distance-aware GATv2 model for spatial transcriptomics cell-cell interaction prediction"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Spatial Cell Interactions Developers" }
]
dependencies = [
  "anndata>=0.10",
  "scanpy>=1.9",
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "umap-learn>=0.5",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "torch>=2.2",
  "torch-geometric>=2.5",
  "pyyaml>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.7",
  "black>=24.0",
  "pre-commit>=3.5",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "C4", "UP", "N", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.isort]
known-first-party = ["spatial_interactions"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
